Enable live updates by adding `parametersChanged` signal to `ParameterPanel`, auto-connecting it in `addRow`, and wiring it to `runSimulation` in `MainWindow`. Replace `Run` button with real-time changes. Refactor `numSteps` to use `size_t`.

+++ gui/src/SimulationControlsPanel.h
@@ -8,5 +8,6 @@ class SimulationControlsPanel : public ParameterPanel {
 public:
     SimulationControlsPanel(QWidget* parent = nullptr, double dt = 0.01, int numSteps = 1000);
     double dt() const;
-    int numSteps() const;
+
+    size_t numSteps() const;
 };
